requirements.txt
Flask==2.3.2
opencv-python-headless
numpy
werkzeug
gunicorn
requests
google-generativeai
scikit-image
supabase
postgrest
